ironclad/requirements.txt
torch
torchvision
facenet-pytorch
Pillow
faiss-cpu==1.9.0
numpy>=1.25.0 
cython>=0.29.21
blosc2
FuzzyTM>=0.4.0  
